Split the project in two and added gradle build
