+++ build.gradle
@@ -1,0 +1,19 @@
+buildscript {
+
+  repositories {
+    jcenter()
+  }
+
+  dependencies {
+    classpath 'com.android.tools.build:gradle:1.3.1'
+  }
+
+}
+
+allprojects {
+  version = '1.0'
+
+    repositories {
+      mavenCentral()
+    }
+}
