+++ example/src/main/AndroidManifest.xml
