+++ example/src/main/res/layout/fragment_main.xml
