+++ example/src/main/res/values/styles.xml
