+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Mon Sep 14 16:24:47 JST 2015
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-2.6-bin.zip
