+++ settings.gradle
@@ -1,0 +1,2 @@
+include 'wanakana-core'
+include 'wanakana-android'
