+++ wanakana-android/build.gradle
@@ -1,0 +1,18 @@
+apply plugin: 'android-library'
+
+android {
+
+  compileSdkVersion 15
+  buildToolsVersion "23.0.1"
+
+  defaultConfig {
+      minSdkVersion 14
+      targetSdkVersion 15
+      versionCode 1
+      versionName "1.0"
+  }
+}
+
+dependencies {
+    compile project(':wanakana-core')
+}
